requirements.txt
yt-dlp
Pillow>=11.2.0
mutagen>=1.47.0
imageio-ffmpeg